Reorganized extensions, consolidated Hashie StringifyKeys implementation.
