+++ lib/hashie/dash.rb
@@ -13,7 +13,8 @@ module Hashie
   # It is preferrable to a Struct because of the in-class
   # API for defining properties as well as per-property defaults.
   class Dash < Hash
-    include PrettyInspect
+    include Hashie::Extensions::PrettyInspect
+
     alias_method :to_s, :inspect
 
     # Defines a property on the Dash. Options are
