+++ lib/hashie/extensions/pretty_inspect.rb
@@ -1,0 +1,19 @@
+module Hashie
+  module Extensions
+    module PrettyInspect
+      def self.included(base)
+        base.send :alias_method, :hash_inspect, :inspect
+        base.send :alias_method, :inspect, :hashie_inspect
+      end
+
+      def hashie_inspect
+        ret = "#<#{self.class}"
+        stringify_keys.keys.sort.each do |key|
+          ret << " #{key}=#{self[key].inspect}"
+        end
+        ret << '>'
+        ret
+      end
+    end
+  end
+end
