+++ spec/hashie/hash_spec.rb
@@ -13,6 +13,12 @@ describe Hash do
     expect(hash).to eq Hashie::Hash['a' => 'hey', '123' => 'bob']
   end
 
+  it '#stringify_keys! turns all keys into strings non-recursively' do
+    hash = Hashie::Hash[:a => 'hey', 123 => { 345 => 'hey' }]
+    hash.stringify_keys!
+    expect(hash).to eq Hashie::Hash['a' => 'hey', '123' => { 345 => 'hey' }]
+  end
+
   it '#stringify_keys returns a hash with stringified keys' do
     hash = Hashie::Hash[:a => 'hey', 123 => 'bob']
     stringified_hash = hash.stringify_keys
